Clean up scripts
Removes testing and coverage-specific code from core/scripts.py.
